// src/shared/models/shared-user.model.ts
import { UserStatus } from 'src/shared/constants/auth.constant'
import z from 'zod'

export const UserSchema = z.object({
  id: z.string(),
  email: z.string().email(),
  name: z.string(),
  password: z.string().min(6).max(100),
  totpSecret: z.string().nullable(),
  phoneNumber: z.string(),
  roleId: z.string(),
  avatar: z.string().nullable(),
  status: z.enum([UserStatus.ACTIVE, UserStatus.INACTIVE, UserStatus.BLOCKED]).default(UserStatus.ACTIVE),
  createdAt: z.date(),
  updatedAt: z.date(),
  deletedAt: z.date().nullable(),

  createdById: z.string().nullable(),
  updatedById: z.string().nullable(),
})

export type UserType = z.infer<typeof UserSchema>